GLStateEGL,GLStateGLX: Change failure to get a "good" visual config t… …o warning
With ea488e9 glmark2 was adjusted to abort if it could not find a
"good" EGL or GLX config.

This commit changes that to instead warn and continue in that case.

Fixes glmark2 running on at least NetBSD-10/amd64 on a ThinkPad
T530, and should also address issue#202 for Ubuntu 22.04 on arm64

Fixes #202